pyproject.toml
[project]
name = "dev-research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "firecrawl-py>=4.3.6",
    "langchain>=0.3.27",
    "langchain-groq>=0.3.8",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
]
